Crate gstreamer_audio

Modules

prelude

Structs

AudioFlags

AudioFormatFlags

AudioFormatInfo
Information for an audio format.

AudioInfo
Information describing audio properties. This information can be filled in from GstCaps with AudioInfo::from_caps.

AudioInfoBuilder

AudioPackFlags

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

StreamVolume
This interface is implemented by elements that provide a stream volume. Examples for such elements are volume and playbin.

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

AudioChannelPosition
Audio channel positions.

AudioEndianness

AudioFormat
Enum value describing the most common audio formats.

AudioLayout
Layout of the audio samples for the different channels.

StreamVolumeFormat
Different representations of a stream volume. StreamVolume::convert_volume allows to convert between the different representations.

Type
A GLib or GLib-based library type

Constants

AUDIO_FORMAT_F32

AUDIO_FORMAT_F64

AUDIO_FORMAT_S16

AUDIO_FORMAT_S18

AUDIO_FORMAT_S20

AUDIO_FORMAT_S24

AUDIO_FORMAT_S32

AUDIO_FORMAT_S2432

AUDIO_FORMAT_U16

AUDIO_FORMAT_U18

AUDIO_FORMAT_U20

AUDIO_FORMAT_U24

AUDIO_FORMAT_U32

AUDIO_FORMAT_U2432

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

StaticType
Types that are supported by GLib dynamic typing.

StreamVolumeExt
Trait containing all StreamVolume methods.

ToValue
Converts to Value.

Functions

audio_buffer_clip

buffer_reorder_channels

get_channel_reorder_map

reorder_channels