Module gstreamer_audio::prelude

Structs

Continue

Continue calling the closure in the future iterations or drop it.

Traits

AllocatorExt

AllocatorExtManual

AudioBaseSinkExt

AudioBaseSrcExt

AudioDecoderExt

AudioDecoderExtManual

AudioEncoderExt

AudioEncoderExtManual

BufferPoolExt

BufferPoolExtManual

Cast

Upcasting and downcasting support.

ChildProxyExt

ChildProxyExtManual

ClockExt

ClockExtManual

CustomTag

DeviceExt

DeviceMonitorExt

DeviceMonitorExtManual

DeviceProviderExt

DeviceProviderExtManual

ElementExt

ElementExtManual

FormattedValue

GObjectExtManualGst

GhostPadExt

GstBinExt

GstBinExtManual

GstObjectExt

GstObjectExtManual

GstParamSpecExt

GstPipelineExtManual

GstPluginExtManual

GstValueExt

IsA

Declares the “is a” relationship.

IsClassFor

Trait for mapping a class struct type to its corresponding instance type.

MessageErrorDomain

MetaAPI

MiniObject

MulDiv

Trait for calculating val * num / denom with different rounding modes and overflow protection.

ObjectExt

ObjectType

Implemented by types representing glib::Object and subclasses of it.

PadExt

PadExtManual

PipelineExt

PluginFeatureExt

PluginFeatureExtManual

PresetExt

ProxyPadExt

ProxyPadExtManual

SpecificFormattedValue

StaticType

Types that are supported by GLib dynamic typing.

StaticVariantType

Returns VariantType of Self.

StreamVolumeExt

SystemClockExt

Tag

TagSetterExt

TagSetterExtManual

ToSendValue

Converts to SendValue.

ToValue

Converts to Value.

ToVariant

Converts to Variant.

TocSetterExt

TypeFindImpl

URIHandlerExt

UnixBusExtManual