Enum gstreamer_audio::AudioChannelPosition

pub enum AudioChannelPosition {
    None,
    Mono,
    Invalid,
    FrontLeft,
    FrontRight,
    FrontCenter,
    Lfe1,
    RearLeft,
    RearRight,
    FrontLeftOfCenter,
    FrontRightOfCenter,
    RearCenter,
    Lfe2,
    SideLeft,
    SideRight,
    TopFrontLeft,
    TopFrontRight,
    TopFrontCenter,
    TopCenter,
    TopRearLeft,
    TopRearRight,
    TopSideLeft,
    TopSideRight,
    TopRearCenter,
    BottomFrontCenter,
    BottomFrontLeft,
    BottomFrontRight,
    WideLeft,
    WideRight,
    SurroundLeft,
    SurroundRight,
    // some variants omitted
}

Audio channel positions.

These are the channels defined in SMPTE 2036-2-2008 Table 1 for 22.2 audio systems with the Surround and Wide channels from DTS Coherent Acoustics (v.1.3.1) and 10.2 and 7.1 layouts. In the caps the actual channel layout is expressed with a channel count and a channel mask, which describes the existing channels. The positions in the bit mask correspond to the enum values. For negotiation it is allowed to have more bits set in the channel mask than the number of channels to specify the allowed channel positions but this is not allowed in negotiated caps. It is not allowed in any situation other than the one mentioned below to have less bits set in the channel mask than the number of channels.

AudioChannelPosition::Mono can only be used with a single mono channel that has no direction information and would be mixed into all directional channels. This is expressed in caps by having a single channel and no channel mask.

AudioChannelPosition::None can only be used if all channels have this position. This is expressed in caps by having a channel mask with no bits set.

As another special case it is allowed to have two channels without a channel mask. This implicitely means that this is a stereo stream with a front left and front right channel.

Variants
NoneMonoInvalidFrontLeftFrontRightFrontCenterLfe1RearLeftRearRightFrontLeftOfCenterFrontRightOfCenterRearCenterLfe2SideLeftSideRightTopFrontLeftTopFrontRightTopFrontCenterTopCenterTopRearLeftTopRearRightTopSideLeftTopSideRightTopRearCenterBottomFrontCenterBottomFrontLeftBottomFrontRightWideLeftWideRightSurroundLeftSurroundRight
Methods

impl AudioChannelPosition

fn to_mask(&self) -> u64

fn positions_to_mask(
positions: &[AudioChannelPosition],
force_order: bool
) -> Option<u64>

fn positions_from_mask(
mask: u64,
positions: &mut [AudioChannelPosition]
) -> Result<(), BoolError>

fn positions_to_valid_order(
positions: &mut [AudioChannelPosition]
) -> Result<(), BoolError>

fn get_fallback_mask(channels: u32) -> u64

fn check_valid_channel_positions(
positions: &[AudioChannelPosition],
force_order: bool
) -> bool